Fix binding when determining version from gemspec
Some gems don't use a namespaced VERSION constant.  When publishing
these gems, gem_publisher would push a tag with it's own version instead
of the gem's version.  This is because the gemspec was evaled within the
GemPublisher module, and therefore the GemPublisher::VERSION constant
was picked up.

Admittedly, these gems aren't well written, but picking up
gem_publisher's version is definitely an error.

eval'ing with the TOPLEVEL_BINDING solves this.