Fix the `-r`/`--repository` option.
For whatever reason, either because it never worked or because a new
version of Rubygems broke backwards compatibility, this switch wasn't
accepting an argument, rather than requiring one. Because there weren't
any tests around this plugin, and we use Rake tasks to do this
ourselves, we didn't spot this until someone tried to use it for real,
resulting in <https://github.com/alphagov/gem_publisher/pull/18>.

There's now enough test coverage around this plugin that we can be more
confident it actually works as intended.