Push a branch as alias of the latest release tag
This is to make diffing master against  whatever the latest released version
of the package is. This allows us to easily see what has been committed to
master but not yet released in a consistent way, eg with a diff like:

https://github.com/alphagov/govuk_frontend_toolkit/compare/latest-release...master

Which means we don't need to know the latest release is tag `v4.10.0` (for example)

For deployed applications we track some alias branches, like `deployed-to-production`
that are set during the deploy jobs, which allow us to do the same thing, and is used
by tools like the deploy-lag-radiator (https://github.com/dsingleton/deploy-lag-radiator)
to visualise when code hasn't been deployed.

Code hanging around has caused us problems in the past, the deploy-lag-radiator helped
a lot for deployed apps, this attempts to do the same for gems.